+++ twitter_clone/main.py
@@ -1,23 +1,47 @@
-try:
-    from .login import login, login_required
-except ValueError:
-    from twitter_clone.login import login, login_required
-
-try:
-    from .feed import other_feed, own_feed, feed
-except ValueError:
-    from twitter_clone.feed import other_feed, own_feed, feed
-
-try:
-    from .modules import app
-except ValueError:
-    from twitter_clone.modules import app
-
 import sqlite3
 from hashlib import md5
 from functools import wraps
+from flask import Flask
 from flask import (g, request, session, redirect, render_template,
                    flash, url_for, abort)
+
+
+app = Flask(__name__)
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    if 'username' in session:
+        return redirect(url_for('home'))
+    cur = g.db.cursor()
+    if request.method == 'POST':
+        user = request.form['username']
+        pw = md5(request.form['password'].encode('utf-8')).hexdigest()
+        cur.execute('SELECT * from user WHERE username = ? AND password = ?',
+                    (user, pw))
+
+        fetched = cur.fetchone()
+        if fetched is None:
+            return "Invalid username or password"
+        fetched = list(fetched)
+        fetched.pop(2)
+
+        check_for = ['user_id','username','fname','lname','bdate']
+        session['logged_in'] = True
+        for idx, check in enumerate(check_for):
+            session[check] = fetched[idx]
+        return redirect(url_for('feed',username=session['username']))
+
+    return render_template('static_templates/layout_login.html')
+
+
+def login_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if 'username' not in session:
+            return redirect(url_for('login', next=request.url))
+        return f(*args, **kwargs)
+    return decorated_function
+
 
 def connect_db(db_name):
     return sqlite3.connect(db_name)
@@ -26,12 +50,43 @@ def connect_db(db_name):
 def before_request():
     g.db = connect_db(app.config['DATABASE'][1])
 
-@app.route('/')
-def home():
-    if 'username' in session:
-        return redirect(url_for('feed',username=session['username']))
+@app.route('/<username>', methods=['GET'])
+def feed(username):
+    if 'username' not in session or session['username'] != username:
+        return other_feed(username)
     else:
-        return redirect(url_for('login'))
+        return own_feed(username)
+
+
+def own_feed(username):
+    user_id = session['user_id']
+    cur = g.db.cursor()
+    cur.execute('SELECT id, content, created from tweet WHERE user_id = ? \
+                ORDER BY id DESC',(user_id,))
+    my_tweets = cur.fetchall()
+    return render_template('static_templates/layout_own_feed.html',
+                            username=username,tweets=my_tweets)
+
+
+def other_feed(username):
+    if 'username' in session:
+        me = session['username']
+    else:
+        me = ''
+    cur = g.db.cursor()
+    cur.execute('SELECT id, username from user WHERE username = ?',
+    (username,))
+    fetched = cur.fetchone()
+    user_id = fetched[0]
+    user = fetched[1]
+
+    cur.execute('SELECT id,content,created from tweet WHERE user_id = ? \
+    ORDER BY id DESC',(user_id,))
+
+    tweets = cur.fetchall()
+    return render_template('static_templates/layout_other_feed.html',
+                            me=me,username=user,tweets=tweets), 200
+
 
 @app.route('/<username>', methods=['POST'])
 def post_tweet(username):
@@ -46,20 +101,28 @@ def post_tweet(username):
     cur.execute('SELECT id, content, created FROM tweet WHERE user_id = ? \
                 ORDER BY id DESC',(user_id,))
     my_tweets = cur.fetchall()
-    flash('You have successfully posted a new Tweet.')
+    flash('Tweet successfully posted.')
     return render_template('static_templates/layout_own_feed.html',username=username,
                             tweets=my_tweets)
+
+
+@app.route('/')
+def home():
+    if 'username' in session:
+        return redirect(url_for('feed',username=session['username']))
+    else:
+        return redirect(url_for('login'))
+
 
 @app.route('/tweets/<int:tweet_id>/delete', methods=['POST'])
 @login_required
 def delete_tweet(tweet_id):
-    if 'username' not in session:
-        abort(403)
     cur = g.db.cursor()
-    user_id = session['user_id']
-    cur.execute('DELETE from tweet WHERE id = ? AND user_id = ?',
-                (tweet_id,user_id))
-    g.db.commit()
+    if 'user_id' in session:
+        user_id = session['user_id']
+        cur.execute('DELETE from tweet WHERE id = ?', (tweet_id,))
+        g.db.commit()
+        print(cur.execute('SELECT * from tweet').fetchall())
     flash('Tweet deleted.')
     return redirect(url_for('home'))
 
@@ -70,7 +133,7 @@ def logout():
     session['logged_in'] = False
     session.pop('username', None)
     session.pop('user_id', None)
-    flash('You have successfully logged out.')
+    flash('Sucessfully logged out.')
     return redirect(url_for('home'))
 
 
@@ -89,16 +152,21 @@ def profile():
 def edit_profile():
     cur = g.db.cursor()
 
-    session['fname'] = request.form['first_name']
-    session['lname'] = request.form['last_name']
-    session['bdate'] = request.form['birth_date']
+    username = request.form['username']
+    fname = request.form['first_name']
+    lname = request.form['last_name']
+    bdate = request.form['birth_date']
 
-    cur.execute('''UPDATE user SET "first_name" = ?, "last_name" = ?,
-                                   "birth_date" = ? WHERE "username" = ?'''
-    ,(session['fname'],session['lname'],session['bdate'],session['username']))
+    session['fname'] = fname
+    session['lname'] = lname
+    session['bdate'] = bdate
+
+    cur.execute('UPDATE user SET "first_name" = ?, "last_name" = ?, "birth_date" = ? WHERE "username" = ?',
+    (fname,lname,bdate,username))
     g.db.commit()
+    flash('Profile successfully updated.')
     return render_template('static_templates/layout_profile.html',
                             username = session['username'],
                             birth_date = session['bdate'],
-                            first_name =session['fname'],
+                            first_name = session['fname'],
                             last_name = session['lname'])
