--- twitter_clone/modules.py
@@ -1,3 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
